+++ test/http/client_test_test.dart
@@ -23,7 +23,7 @@ void main() {
 }
 
 class TestController extends RESTController {
-  @Bind.get()
+  @Operation.get()
   Future<Response> get() async {
     return new Response.ok([
       {"id": 1},